requirements.txt
appdirs==1.4.0
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
blinker==1.4
click==6.7
decorator==4.0.11
dominate==2.3.1
enum34==1.1.6
Flask==0.12
Flask-Login==0.3.2
Flask-Mail==0.9.1
flask-nav==0.6
Flask-Navigation==0.2.0
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
google-api-python-client==1.6.2
httplib2==0.10.3
ipython==5.2.2
ipython-genutils==0.1.0
isbnlib==3.6.6
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
oauth2client==4.0.0
packaging==16.8
passlib==1.7.1
pathlib2==2.2.1
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.2
pyasn1-modules==0.0.8
Pygments==2.2.0
PyMySQL==0.7.9
pyparsing==2.1.10
rsa==3.4.2
scandir==1.4
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.5
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
traitlets==4.3.1
uritemplate==3.0.0
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.11.15
WTForms==2.1
